Refactors TKEVD to CATKEVD, plus calculate_diffusivities refactor
