+++ src/Models/HydrostaticFreeSurfaceModels/hydrostatic_free_surface_tendency_kernel_functions.jl
@@ -6,8 +6,9 @@ using Oceananigans.StokesDrift
 using Oceananigans.TurbulenceClosures: ∂ⱼ_τ₁ⱼ, ∂ⱼ_τ₂ⱼ, ∇_dot_qᶜ
 using Oceananigans.Advection: div_Uc
 
-using Oceananigans.TurbulenceClosures: shear_production, buoyancy_flux, dissipation, TKETracerIndex
-import Oceananigans.TurbulenceClosures: hydrostatic_turbulent_kinetic_energy_tendency
+using Oceananigans.TurbulenceClosures.CATKEVerticalDiffusivities: shear_production, buoyancy_flux, dissipation, TKETracerIndex
+
+import Oceananigans.TurbulenceClosures.CATKEVerticalDiffusivities: hydrostatic_turbulent_kinetic_energy_tendency
 
 """
 Return the tendency for the horizontal velocity in the x-direction, or the east-west 
@@ -97,6 +98,7 @@ where `c = C[tracer_index]`.
                                                           velocities,
                                                           free_surface,
                                                           tracers,
+                                                          top_tracer_bcs,
                                                           diffusivities,
                                                           auxiliary_fields,
                                                           forcing,
@@ -141,6 +143,7 @@ end
                                                                velocities,
                                                                free_surface,
                                                                tracers,
+                                                               top_tracer_bcs,
                                                                diffusivities,
                                                                auxiliary_fields,
                                                                forcing,
@@ -155,7 +158,7 @@ end
              - ∇_dot_qᶜ(i, j, k, grid, closure, e, tke_index, clock, diffusivities, tracers, buoyancy, velocities)
              + shear_production(i, j, k, grid, closure, clock, velocities, tracers, buoyancy, diffusivities)
              + buoyancy_flux(i, j, k, grid, closure, velocities, tracers, buoyancy)
-             - dissipation(i, j, k, grid, closure, tracers, buoyancy)
+             - dissipation(i, j, k, grid, closure, velocities, tracers, buoyancy, clock, tracer_bcs)
              + forcing(i, j, k, grid, clock, hydrostatic_prognostic_fields(velocities, free_surface, tracers)))
 end
