+++ src/TurbulenceClosures/TurbulenceClosures.jl
@@ -73,8 +73,10 @@ include("turbulence_closure_implementations/anisotropic_biharmonic_diffusivity.j
 include("turbulence_closure_implementations/leith_enstrophy_diffusivity.jl")
 include("turbulence_closure_implementations/smagorinsky_lilly.jl")
 include("turbulence_closure_implementations/anisotropic_minimum_dissipation.jl")
-include("turbulence_closure_implementations/tke_based_vertical_diffusivity.jl")
+include("turbulence_closure_implementations/CATKEVerticalDiffusivities/CATKEVerticalDiffusivities.jl")
 include("turbulence_closure_implementations/convective_adjustment_vertical_diffusivity.jl")
+
+using .CATKEVerticalDiffusivities: CATKEVerticalDiffusivity
 
 # Miscellaneous utilities
 include("diffusivity_fields.jl")
