+++ src/TurbulenceClosures/turbulence_closure_implementations/nothing_closure.jl
@@ -3,4 +3,5 @@
 @inline ∂ⱼ_τ₂ⱼ(i, j, k, grid::AbstractGrid{FT}, ::Nothing, args...) where FT = zero(FT)
 @inline ∂ⱼ_τ₃ⱼ(i, j, k, grid::AbstractGrid{FT}, ::Nothing, args...) where FT = zero(FT)
 
-calculate_diffusivities!(K, arch, grid, closure::Nothing, args...) = nothing
+calculate_diffusivities!(diffusivities, ::Nothing, args...) = nothing
+calculate_diffusivities!(::Nothing, ::Nothing, args...) = nothing
